pyproject.toml
[project]
name = "customer-churn-etl"
version = "0.1.0"
description = "ETL pipeline para el análisis y reducción de churn de clientes en empresas de telecomunicaciones. El proyecto incluye la extracción de datos desde archivos CSV, transformación para limpieza y enriquecimiento de datos, carga a una base de datos relacional y la creación de un dashboard para visualizar métricas clave y tendencias de churn."
authors = [
    {name = "JMTrujilloC",email = "juan_man.trujillo@uao.edu.co"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "kaggle (>=1.6.17,<2.0.0)",
    "sqlalchemy (==1.4.36)",
    "apache-airflow (>=2.10.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
